+++ src/vibe3/runtime/heartbeat.py
@@ -14,9 +14,10 @@ from vibe3.orchestra.logging import (
     append_orchestra_event,
     append_orchestra_run_separator,
 )
-from vibe3.runtime.periodic_check_executor import execute_periodic_check
-from vibe3.runtime.service_protocol import ServiceBase
 from vibe3.services.error_tracking_service import ErrorTrackingService
+
+from .periodic_check_executor import execute_periodic_check
+from .service_protocol import ServiceBase
 
 if TYPE_CHECKING:
     from vibe3.domain.failed_gate import FailedGate
